fix: rename file name (dat -> day button)
